Removing option to leave running clusters
no_cleanup used to leave the clusters up and running, but that's really
not necessary anymore. no_cleanup would be more useful if it skipped
the bucket cleanup and still did the download.

+++ cluster/src/main/java/com/hartwig/pipeline/io/RuntimeBucket.java
@@ -42,11 +42,13 @@ public class RuntimeBucket {
     }
 
     public void cleanup() {
-        for (Blob blob : bucket.list().iterateAll()) {
-            blob.delete();
+        if (bucket.exists()) {
+            for (Blob blob : bucket.list().iterateAll()) {
+                blob.delete();
+            }
+            bucket.delete();
+            LOGGER.info("Cleaned up all data in runtime bucket [{}]", bucket.getName());
         }
-        bucket.delete();
-        LOGGER.info("Cleaned up all data in runtime bucket [{}]", bucket.getName());
     }
 
     public String getName() {
